Fix image originals in quick preivew (#1889)
* render original images

* separate original component + error boundary

* types
