+++ interface/app/$libraryId/Explorer/FilePath/Thumb.tsx
@@ -1,29 +1,19 @@
-import { getIcon, getIconByName, iconNames } from '@sd/assets/util';
+import { getIcon, getIconByName } from '@sd/assets/util';
 import clsx from 'clsx';
-import {
-	memo,
-	SyntheticEvent,
-	useEffect,
-	useLayoutEffect,
-	useMemo,
-	useRef,
-	useState,
-	type CSSProperties,
-	type ImgHTMLAttributes,
-	type RefObject,
-	type VideoHTMLAttributes
-} from 'react';
+import { memo, SyntheticEvent, useMemo, useRef, useState, type ImgHTMLAttributes } from 'react';
+import { ErrorBoundary } from 'react-error-boundary';
 import { getItemFilePath, useLibraryContext, type ExplorerItem } from '@sd/client';
-import { PDFViewer, TextViewer } from '~/components';
-import { useCallbackToWatchResize, useIsDark } from '~/hooks';
+import { useIsDark } from '~/hooks';
 import { pdfViewerEnabled } from '~/util/pdfViewer';
 import { usePlatform } from '~/util/Platform';
 
 import { useExplorerContext } from '../Context';
-import { getExplorerStore } from '../store';
-import { ExplorerItemData, useExplorerItemData } from '../util';
+import { useExplorerItemData } from '../util';
+import { Image, ImageProps } from './Image';
 import LayeredFileIcon from './LayeredFileIcon';
+import { Original } from './Original';
 import classes from './Thumb.module.scss';
+import { useBlackBars, useSize } from './utils';
 
 export interface ThumbProps {
 	data: ExplorerItem;
@@ -44,10 +34,7 @@ export interface ThumbProps {
 	isSidebarPreview?: boolean;
 }
 
-type ThumbType =
-	| { variant: 'original'; renderer: OriginalRenderer }
-	| { variant: 'thumbnail' }
-	| { variant: 'icon' };
+type ThumbType = { variant: 'original' } | { variant: 'thumbnail' } | { variant: 'icon' };
 
 export const FileThumb = memo((props: ThumbProps) => {
 	const isDark = useIsDark();
@@ -71,18 +58,7 @@ export const FileThumb = memo((props: ThumbProps) => {
 	);
 
 	const thumbType = useMemo<ThumbType>(() => {
-		let thumbType = props.loadOriginal ? 'original' : 'thumbnail';
-
-		if (thumbType === 'original') {
-			if (loadState.original !== 'error') {
-				const kind = originalRendererKind(itemData);
-				const renderer = ORIGINAL_RENDERERS[kind];
-
-				if (renderer) return { variant: 'original', renderer };
-			}
-
-			thumbType = 'thumbnail';
-		}
+		const thumbType = 'thumbnail';
 
 		if (thumbType === 'thumbnail')
 			if (
@@ -93,7 +69,7 @@ export const FileThumb = memo((props: ThumbProps) => {
 				return { variant: 'thumbnail' };
 
 		return { variant: 'icon' };
-	}, [props.loadOriginal, itemData, loadState]);
+	}, [itemData, loadState]);
 
 	const src = useMemo(() => {
 		const locationId =
@@ -149,6 +125,65 @@ export const FileThumb = memo((props: ThumbProps) => {
 		);
 	};
 
+	const _childClassName =
+		typeof props.childClassName === 'function'
+			? props.childClassName(thumbType)
+			: props.childClassName;
+
+	const className = clsx(childClassName, _childClassName);
+
+	const thumbnail = (() => {
+		if (!src) return null;
+
+		switch (thumbType.variant) {
+			case 'thumbnail':
+				return (
+					<Thumbnail
+						src={src}
+						cover={props.cover}
+						onLoad={() => onLoad('thumbnail')}
+						onError={(e) => onError('thumbnail', e)}
+						decoding={props.size ? 'async' : 'sync'}
+						className={clsx(
+							props.cover
+								? [
+										'min-h-full min-w-full object-cover object-center',
+										_childClassName
+								  ]
+								: className,
+							props.frame && !(itemData.kind === 'Video' && props.blackBars)
+								? frameClassName
+								: null
+						)}
+						crossOrigin="anonymous" // Here it is ok, because it is not a react attr
+						blackBars={props.blackBars && itemData.kind === 'Video' && !props.cover}
+						blackBarsSize={props.blackBarsSize}
+						extension={
+							props.extension && itemData.extension && itemData.kind === 'Video'
+								? itemData.extension
+								: undefined
+						}
+					/>
+				);
+
+			case 'icon':
+				return (
+					<LayeredFileIcon
+						src={src}
+						kind={itemData.kind}
+						extension={itemData.extension}
+						onLoad={() => onLoad('icon')}
+						onError={(e) => onError('icon', e)}
+						decoding={props.size ? 'async' : 'sync'}
+						className={className}
+						draggable={false}
+					/>
+				);
+			default:
+				return null;
+		}
+	})();
+
 	return (
 		<div
 			style={{
@@ -164,91 +199,34 @@ export const FileThumb = memo((props: ThumbProps) => {
 				props.className
 			)}
 		>
-			{(() => {
-				if (!src) return;
-
-				const _childClassName =
-					typeof props.childClassName === 'function'
-						? props.childClassName(thumbType)
-						: props.childClassName;
-
-				const className = clsx(childClassName, _childClassName);
-
-				switch (thumbType.variant) {
-					case 'original':
-						return thumbType.renderer({
-							src,
-							className,
-							frameClassName,
-							itemData,
-							isDark,
-							childClassName,
-							onLoad: () => onLoad('original'),
-							onError: (e) => onError('original', e),
-							size: props.size,
-							mediaControls: props.mediaControls,
-							frame: props.frame,
-							isSidebarPreview: props.isSidebarPreview,
-							pauseVideo: props.pauseVideo,
-							blackBars: props.blackBars,
-							blackBarsSize: props.blackBarsSize
-						});
-
-					// eslint-disable-next-line no-fallthrough
-					case 'thumbnail':
-						return (
-							<Thumbnail
-								src={src}
-								cover={props.cover}
-								onLoad={() => onLoad('thumbnail')}
-								onError={(e) => onError('thumbnail', e)}
-								decoding={props.size ? 'async' : 'sync'}
-								className={clsx(
-									props.cover
-										? [
-												'min-h-full min-w-full object-cover object-center',
-												_childClassName
-										  ]
-										: className,
-									props.frame && !(itemData.kind === 'Video' && props.blackBars)
-										? frameClassName
-										: null
-								)}
-								crossOrigin="anonymous" // Here it is ok, because it is not a react attr
-								blackBars={
-									props.blackBars && itemData.kind === 'Video' && !props.cover
-								}
-								blackBarsSize={props.blackBarsSize}
-								extension={
-									props.extension &&
-									itemData.extension &&
-									itemData.kind === 'Video'
-										? itemData.extension
-										: undefined
-								}
-							/>
-						);
-
-					case 'icon':
-						return (
-							<LayeredFileIcon
-								src={src}
-								kind={itemData.kind}
-								extension={itemData.extension}
-								onLoad={() => onLoad('icon')}
-								onError={(e) => onError('icon', e)}
-								decoding={props.size ? 'async' : 'sync'}
-								className={className}
-								draggable={false}
-							/>
-						);
-				}
-			})()}
+			{props.loadOriginal ? (
+				<ErrorBoundary fallback={thumbnail}>
+					<Original
+						onLoad={() => onLoad('original')}
+						onError={(e) => onError('original', e)}
+						filePath={filePath}
+						className={className}
+						frameClassName={frameClassName}
+						itemData={itemData}
+						isDark={isDark}
+						childClassName={childClassName}
+						size={props.size}
+						mediaControls={props.mediaControls}
+						frame={props.frame}
+						isSidebarPreview={props.isSidebarPreview}
+						pauseVideo={props.pauseVideo}
+						blackBars={props.blackBars}
+						blackBarsSize={props.blackBarsSize}
+					/>
+				</ErrorBoundary>
+			) : (
+				thumbnail
+			)}
 		</div>
 	);
 });
 
-interface ThumbnailProps extends ImgHTMLAttributes<HTMLImageElement> {
+interface ThumbnailProps extends Omit<ImageProps, 'blackBarsStyle' | 'size'> {
 	cover?: boolean;
 	blackBars?: boolean;
 	blackBarsSize?: number;
@@ -273,17 +251,19 @@ const Thumbnail = memo(
 
 		return (
 			<>
-				<img
-					// Order matter for crossOrigin attr
-					// https://github.com/facebook/react/issues/14035#issuecomment-642227899
-					{...(crossOrigin ? { crossOrigin } : {})}
-					ref={ref}
-					draggable={false}
-					style={{ ...(blackBars ? blackBarsStyle : {}) }}
-					className={clsx(blackBars && size.width === 0 && 'invisible', className)}
+				<Image
 					{...props}
+					{...{
+						className: clsx(
+							className,
+							blackBarsStyle && size.width === 0 && 'invisible'
+						),
+						cover,
+						style: blackBars ? blackBarsStyle : undefined,
+						size,
+						ref
+					}}
 				/>
-
 				{(cover || (size && size.width > 80)) && extension && (
 					<div
 						style={{
@@ -307,189 +287,3 @@ const Thumbnail = memo(
 		);
 	}
 );
-
-interface OriginalRendererProps {
-	src: string;
-	className: string;
-	frameClassName: string;
-	itemData: ExplorerItemData;
-	isDark: boolean;
-	childClassName?: string;
-	size?: number;
-	mediaControls?: boolean;
-	frame?: boolean;
-	isSidebarPreview?: boolean;
-	pauseVideo?: boolean;
-	blackBars?: boolean;
-	blackBarsSize?: number;
-	onLoad?(): void;
-	onError?(e: ErrorEvent | SyntheticEvent<Element, Event>): void;
-}
-
-const TEXT_RENDERER: OriginalRenderer = (props) => (
-	<TextViewer
-		src={props.src}
-		onLoad={props.onLoad}
-		onError={props.onError}
-		className={clsx(
-			'textviewer-scroll h-full w-full overflow-y-auto whitespace-pre-wrap break-words px-4 font-mono',
-			!props.mediaControls ? 'overflow-hidden' : 'overflow-auto',
-			props.className,
-			props.frame && [props.frameClassName, '!bg-none p-2']
-		)}
-		codeExtension={
-			((props.itemData.kind === 'Code' || props.itemData.kind === 'Config') &&
-				props.itemData.extension) ||
-			''
-		}
-		isSidebarPreview={props.isSidebarPreview}
-	/>
-);
-
-type OriginalRenderer = (props: OriginalRendererProps) => JSX.Element;
-
-function originalRendererKind(itemData: ExplorerItemData) {
-	return itemData.extension === 'pdf' ? 'PDF' : itemData.kind;
-}
-
-type OriginalRendererKind = ReturnType<typeof originalRendererKind>;
-
-const ORIGINAL_RENDERERS: {
-	[K in OriginalRendererKind]?: OriginalRenderer;
-} = {
-	PDF: (props) => (
-		<PDFViewer
-			src={props.src}
-			onLoad={props.onLoad}
-			onError={props.onError}
-			className={clsx('h-full w-full', props.className, props.frame && props.frameClassName)}
-			crossOrigin="anonymous" // Here it is ok, because it is not a react attr
-		/>
-	),
-	Text: TEXT_RENDERER,
-	Code: TEXT_RENDERER,
-	Config: TEXT_RENDERER,
-	Video: (props) => (
-		<Video
-			src={props.src}
-			onLoadedData={props.onLoad}
-			onError={props.onError}
-			paused={props.pauseVideo}
-			controls={props.mediaControls}
-			blackBars={props.blackBars}
-			blackBarsSize={props.blackBarsSize}
-			className={clsx(
-				props.className,
-				props.frame && !props.blackBars && props.frameClassName
-			)}
-		/>
-	),
-	Audio: (props) => (
-		<>
-			<img
-				src={getIcon(iconNames.Audio, props.isDark, props.itemData.extension)}
-				onLoad={props.onLoad}
-				decoding={props.size ? 'async' : 'sync'}
-				className={props.childClassName}
-				draggable={false}
-			/>
-			{props.mediaControls && (
-				<audio
-					// Order matter for crossOrigin attr
-					crossOrigin="anonymous"
-					src={props.src}
-					onError={props.onError}
-					controls
-					autoPlay
-					className="absolute left-2/4 top-full w-full -translate-x-1/2 translate-y-[-150%]"
-				>
-					<p>Audio preview is not supported.</p>
-				</audio>
-			)}
-		</>
-	)
-};
-
-interface VideoProps extends VideoHTMLAttributes<HTMLVideoElement> {
-	paused?: boolean;
-	blackBars?: boolean;
-	blackBarsSize?: number;
-}
-
-const Video = memo(({ paused, blackBars, blackBarsSize, className, ...props }: VideoProps) => {
-	const ref = useRef<HTMLVideoElement>(null);
-
-	const size = useSize(ref);
-	const { style: blackBarsStyle } = useBlackBars(size, blackBarsSize);
-
-	useEffect(() => {
-		if (!ref.current) return;
-		paused ? ref.current.pause() : ref.current.play();
-	}, [paused]);
-
-	return (
-		<video
-			// Order matter for crossOrigin attr
-			crossOrigin="anonymous"
-			ref={ref}
-			autoPlay={!paused}
-			onVolumeChange={(e) => {
-				const video = e.target as HTMLVideoElement;
-				getExplorerStore().mediaPlayerVolume = video.volume;
-			}}
-			onCanPlay={(e) => {
-				const video = e.target as HTMLVideoElement;
-				// Why not use the element's attribute? Because React...
-				// https://github.com/facebook/react/issues/10389
-				video.loop = !props.controls;
-				video.muted = !props.controls;
-				video.volume = getExplorerStore().mediaPlayerVolume;
-			}}
-			playsInline
-			draggable={false}
-			style={{ ...(blackBars ? blackBarsStyle : {}) }}
-			className={clsx(blackBars && size.width === 0 && 'invisible', className)}
-			{...props}
-		>
-			<p>Video preview is not supported.</p>
-		</video>
-	);
-});
-
-const useSize = (ref: RefObject<Element>) => {
-	const [size, setSize] = useState({ width: 0, height: 0 });
-
-	useCallbackToWatchResize(({ width, height }) => setSize({ width, height }), [], ref);
-
-	return size;
-};
-
-const useBlackBars = (videoSize: { width: number; height: number }, blackBarsSize?: number) => {
-	return useMemo(() => {
-		const { width, height } = videoSize;
-
-		const orientation = height > width ? 'vertical' : 'horizontal';
-
-		const barSize =
-			blackBarsSize ||
-			Math.floor(Math.ceil(orientation === 'vertical' ? height : width) / 10);
-
-		const xBarSize = orientation === 'vertical' ? barSize : 0;
-		const yBarSize = orientation === 'horizontal' ? barSize : 0;
-
-		return {
-			size: {
-				x: xBarSize,
-				y: yBarSize
-			},
-			style: {
-				borderLeftWidth: xBarSize,
-				borderRightWidth: xBarSize,
-				borderTopWidth: yBarSize,
-				borderBottomWidth: yBarSize,
-				borderColor: 'black',
-				borderRadius: 4
-			} satisfies CSSProperties
-		};
-	}, [videoSize, blackBarsSize]);
-};
