+++ interface/app/$libraryId/Explorer/QuickPreview/index.tsx
@@ -20,6 +20,7 @@ import {
 	useRef,
 	useState
 } from 'react';
+import { ErrorBoundary } from 'react-error-boundary';
 import { useKey } from 'rooks';
 import {
 	ExplorerItem,
@@ -46,6 +47,7 @@ import ExplorerContextMenu, {
 	SharedItems
 } from '../ContextMenu';
 import { Conditional } from '../ContextMenu/ConditionalItem';
+import { Original } from '../FilePath/Original';
 import { FileThumb } from '../FilePath/Thumb';
 import { SingleItemMetadata } from '../Inspector';
 import { ImageSlider } from './ImageSlider';
